24_10_15(final)

+++ src/main/java/com/green/day05/ch06/MethodReturns.java
@@ -7,7 +7,7 @@ public class MethodReturns {
         result = adder(n1, n2); // 더한값을 리턴 9
         System.out.printf("%d + %d = %d\n",n1, n2, result);
         System.out.printf("%.1f x %.1f = %.2f\n", 3.5, 3.5, square(3.5f));
-        System.out.println((int)square(6));
+        System.out.println(square(6));
         System.out.println(square(4.5f));
     }
     static int adder(int n1, int n2) {
